+++ src/main/java/domain/Hand.java
@@ -1,39 +1,23 @@
 package domain;
 
 import domain.card.Card;
-import java.util.ArrayList;
+import domain.card.Cards;
 import java.util.List;
 
 public class Hand {
     public static final int BLACKJACK_SCORE = 21;
-    private final List<Card> cards;
+    private final Cards cards;
 
     public Hand() {
-        cards = new ArrayList<>();
+        cards = new Cards();
+    }
+
+    public Hand(Cards cards) {
+        this.cards = cards;
     }
 
     public int getSum() {
-        int sum = getInitSum();
-        long aceCount = countAce();
-
-        while (aceCount > 0 && sum > BLACKJACK_SCORE) {
-            sum -= 10;
-            aceCount--;
-        }
-
-        return sum;
-    }
-
-    private int getInitSum() {
-        return cards.stream()
-                .mapToInt(Card::getScore)
-                .sum();
-    }
-
-    private long countAce() {
-        return cards.stream()
-                .filter(Card::isAce)
-                .count();
+        return cards.getSum();
     }
 
     public boolean isBust() {
@@ -44,15 +28,19 @@ public class Hand {
         return getSum() == BLACKJACK_SCORE;
     }
 
-    public void addCard(Card card) {
-        cards.add(card);
+    public Hand addCard(Card card) {
+        return new Hand(cards.addCard(card));
     }
 
-    public List<Card> getCards() {
+    public Cards getCards() {
         return cards;
     }
 
-    public Card getFirst() {
+    public List<Card> getCardsList() {
+        return cards.cards();
+    }
+
+    public Card getFirstCard() {
         return cards.getFirst();
     }
