+++ src/main/java/domain/card/Cards.java
@@ -1,5 +1,7 @@
 package domain.card;
 
+import static domain.Hand.BLACKJACK_SCORE;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,4 +20,37 @@ public record Cards(List<Card> cards) {
         cards.add(card);
         return new Cards(cards);
     }
+
+    public int getSum() {
+        int sum = getInitSum();
+        long aceCount = countAce();
+
+        while (aceCount > 0 && sum > BLACKJACK_SCORE) {
+            sum -= 10;
+            aceCount--;
+        }
+
+        return sum;
+    }
+
+    public Card getFirst() {
+        return cards.getFirst();
+    }
+
+    public int size() {
+        return cards.size();
+    }
+
+    private int getInitSum() {
+        return cards.stream()
+                .mapToInt(Card::getScore)
+                .sum();
+    }
+
+    private long countAce() {
+        return cards.stream()
+                .filter(Card::isAce)
+                .count();
+    }
+
 }
