+++ src/main/java/domain/participant/Dealer.java
@@ -5,6 +5,6 @@ import domain.card.Card;
 public class Dealer extends Participant {
 
     public Card getFirstCard() {
-        return hand.getFirst();
+        return hand.getFirstCard();
     }
 }
